flask_webinar/app/requirements.txt
flask
flask_sqlalchemy
flask-marshmallow
marshmallow_sqlalchemy
psycopg2-binary
